api/pyproject.toml
[tool.poetry]
name = "youtube-downloader-api"
version = "0.1.0"
description = ""
authors = ["lipowskm <m.lipowski95@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
rq = "^1.13.0"
unidecode = "^1.3.6"
dynaconf = "^3.1.12"
pytube = "^15.0.0"
rq-scheduler = "^0.13.1"
rq-dashboard = "^0.6.7"
tldextract = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
httpx = "^0.23.3"
fakeredis = "^2.10.2"
pytest-mock = "^3.10.0"
types-redis = "^4.6.0.7"

[tool.poetry.group.nox.dependencies]
nox-poetry = "*"

[tool.poetry.group.type_check.dependencies]
mypy = "*"
types-requests = "^2.28.2"

[tool.poetry.group.lint.dependencies]
ruff = "*"

[tool.poetry.group.fmt.dependencies]
black = "*"
isort = "*"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88
force_sort_within_sections = true
src_paths = ["torrent_player"]
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.ruff]
line-length = 88
select = ["E", "F"]
ignore = []
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"